commit formation+user 3/4 (lien back))

+++ src/Controller/CategorieFormationController.php
@@ -5,6 +5,7 @@ namespace App\Controller;
 use App\Entity\CategorieFormation;
 use App\Form\CategorieFormationType;
 use App\Repository\CategorieFormationRepository;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -13,17 +14,27 @@ use Symfony\Component\Routing\Annotation\Route;
 #[Route('/categorie/formation')]
 class CategorieFormationController extends AbstractController
 {
-    #[Route('/', name: 'app_categorie_formation_index', methods: ['GET'])]
+    #[Route('/', name: 'app_categorie_formation_index', methods: ['GET']), IsGranted('ROLE_ADMIN')]
     public function index(CategorieFormationRepository $categorieFormationRepository): Response
     {
+        $user = $this->getUser();
+
+        if (!$user) {
+            return $this->redirectToRoute('app_login');
+        }
         return $this->render('categorie_formation/index.html.twig', [
             'categorie_formations' => $categorieFormationRepository->findAll(),
         ]);
     }
 
-    #[Route('/new', name: 'app_categorie_formation_new', methods: ['GET', 'POST'])]
+    #[Route('/new', name: 'app_categorie_formation_new', methods: ['GET', 'POST']), IsGranted('ROLE_ADMIN')]
     public function new(Request $request, CategorieFormationRepository $categorieFormationRepository): Response
     {
+        $user = $this->getUser();
+
+        if (!$user) {
+            return $this->redirectToRoute('app_login');
+        }
         $categorieFormation = new CategorieFormation();
         $form = $this->createForm(CategorieFormationType::class, $categorieFormation);
         $form->handleRequest($request);
@@ -40,17 +51,27 @@ class CategorieFormationController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}', name: 'app_categorie_formation_show', methods: ['GET'])]
+    #[Route('/{id}', name: 'app_categorie_formation_show', methods: ['GET']), IsGranted('ROLE_ADMIN')]
     public function show(CategorieFormation $categorieFormation): Response
     {
+        $user = $this->getUser();
+
+        if (!$user) {
+            return $this->redirectToRoute('app_login');
+        }
         return $this->render('categorie_formation/show.html.twig', [
             'categorie_formation' => $categorieFormation,
         ]);
     }
 
-    #[Route('/{id}/edit', name: 'app_categorie_formation_edit', methods: ['GET', 'POST'])]
+    #[Route('/{id}/edit', name: 'app_categorie_formation_edit', methods: ['GET', 'POST']), IsGranted('ROLE_ADMIN')]
     public function edit(Request $request, CategorieFormation $categorieFormation, CategorieFormationRepository $categorieFormationRepository): Response
     {
+        $user = $this->getUser();
+
+        if (!$user) {
+            return $this->redirectToRoute('app_login');
+        }
         $form = $this->createForm(CategorieFormationType::class, $categorieFormation);
         $form->handleRequest($request);
 
@@ -66,9 +87,14 @@ class CategorieFormationController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}', name: 'app_categorie_formation_delete', methods: ['POST'])]
+    #[Route('/{id}', name: 'app_categorie_formation_delete', methods: ['POST']), IsGranted('ROLE_ADMIN')]
     public function delete(Request $request, CategorieFormation $categorieFormation, CategorieFormationRepository $categorieFormationRepository): Response
     {
+        $user = $this->getUser();
+
+        if (!$user) {
+            return $this->redirectToRoute('app_login');
+        }
         if ($this->isCsrfTokenValid('delete'.$categorieFormation->getId(), $request->request->get('_token'))) {
             $categorieFormationRepository->remove($categorieFormation, true);
         }
