+++ src/Controller/FormationController.php
@@ -5,6 +5,7 @@ namespace App\Controller;
 use App\Entity\Formation;
 use App\Form\FormationType;
 use App\Repository\FormationRepository;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\File\Exception\FileException;
 use Symfony\Component\HttpFoundation\Request;
@@ -15,23 +16,34 @@ use Symfony\Component\String\Slugger\SluggerInterface;
 #[Route('/formation')]
 class FormationController extends AbstractController
 {
-    #[Route('/', name: 'app_formation_index', methods: ['GET'])]
+    #[Route('/', name: 'app_formation_index', methods: ['GET']), IsGranted('ROLE_ADMIN')]
     public function index(FormationRepository $formationRepository): Response
     {
+        $user = $this->getUser();
+
+        if (!$user) {
+            return $this->redirectToRoute('app_login');
+        }
         return $this->render('formation/index.html.twig', [
-            'formations' => $formationRepository->findAll(),
+            'user' => $this->getUser(),
+            'formations' => $formationRepository->findAll()
         ]);
     }
 
     #[Route('/front', name: 'app_formation_index_front', methods: ['GET'])]
     public function indexfront(FormationRepository $formationRepository): Response
     {
+        $user = $this->getUser();
+
+        if (!$user) {
+            return $this->redirectToRoute('app_login');
+        }
         return $this->render('formation/index_front.html.twig', [
             'formations' => $formationRepository->findAll(),
         ]);
     }
 
-    #[Route('/new', name: 'app_formation_new', methods: ['GET', 'POST'])]
+    #[Route('/new', name: 'app_formation_new', methods: ['GET', 'POST']),IsGranted('ROLE_ADMIN')]
     public function new(Request $request, FormationRepository $formationRepository , SluggerInterface $slugger): Response
     {
         $formation = new Formation();
@@ -51,7 +63,7 @@ class FormationController extends AbstractController
 
                 try {
                     $pictureFile->move(
-                        $this->getParameter('images'),
+                        $this->getParameter('users_directory'),
                         $newFilename
                     );
                 } catch (FileException $e) {
@@ -72,9 +84,14 @@ class FormationController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}', name: 'app_formation_show', methods: ['GET'])]
+    #[Route('/{id}', name: 'app_formation_show', methods: ['GET']),IsGranted('ROLE_ADMIN')]
     public function show(Formation $formation): Response
     {
+        $user = $this->getUser();
+
+        if (!$user) {
+            return $this->redirectToRoute('app_login');
+        }
         return $this->render('formation/show.html.twig', [
             'formation' => $formation,
         ]);
@@ -83,14 +100,24 @@ class FormationController extends AbstractController
     #[Route('/front/{id}', name: 'app_formation_show_front', methods: ['GET'])]
     public function show_front(Formation $formation): Response
     {
+        $user = $this->getUser();
+
+        if (!$user) {
+            return $this->redirectToRoute('app_login');
+        }
         return $this->render('formation/show_front.html.twig', [
             'formation' => $formation,
         ]);
     }
 
-    #[Route('/{id}/edit', name: 'app_formation_edit', methods: ['GET', 'POST'])]
+    #[Route('/{id}/edit', name: 'app_formation_edit', methods: ['GET', 'POST']),IsGranted('ROLE_ADMIN')]
     public function edit(Request $request, Formation $formation, FormationRepository $formationRepository,SluggerInterface $slugger): Response
     {
+        $user = $this->getUser();
+
+        if (!$user) {
+            return $this->redirectToRoute('app_login');
+        }
         $form = $this->createForm(FormationType::class, $formation);
         $form->handleRequest($request);
 
@@ -103,7 +130,7 @@ class FormationController extends AbstractController
 
                 try {
                     $pictureFile->move(
-                        $this->getParameter('images'),
+                        $this->getParameter('users_directory'),
                         $newFilename
                     );
                 } catch (FileException $e) {
@@ -123,9 +150,14 @@ class FormationController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}', name: 'app_formation_delete', methods: ['POST'])]
+    #[Route('/{id}', name: 'app_formation_delete', methods: ['POST']),IsGranted('ROLE_ADMIN')]
     public function delete(Request $request, Formation $formation, FormationRepository $formationRepository): Response
     {
+        $user = $this->getUser();
+
+        if (!$user) {
+            return $this->redirectToRoute('app_login');
+        }
         if ($this->isCsrfTokenValid('delete'.$formation->getId(), $request->request->get('_token'))) {
             $formationRepository->remove($formation, true);
         }
@@ -145,7 +177,11 @@ class FormationController extends AbstractController
     }
     #[Route('/front/bycentre/{id}', name: 'app_formation_bycentre')]
     public function getByCentre($id,FormationRepository $repo) : Response {
+        $user = $this->getUser();
 
+        if (!$user) {
+            return $this->redirectToRoute('app_login');
+        }
         $formation = $repo->getBycentre($id);
         return $this->renderForm('formation\index_front_bycentre.html.twig', [
             'formations' => $formation,
